refac: split dtos into separate folders
