+++ backend/src/chat/dto/chatroom/update-chatroom-password.dto.ts
@@ -1,6 +1,6 @@
 import { IsNotEmpty, IsString, IsNumber } from 'class-validator';
 
-export class updatePasswordDto {
+export class UpdateChatroomPasswordDto {
   @IsNumber()
   @IsNotEmpty()
   chatroomId: number;
