--- backend/src/chat/dto/on-get-rooms.dto.ts
@@ -1,7 +1,0 @@
-import { IsNotEmpty, IsNumber } from 'class-validator';
-
-export class OnGetRoomsDto {
-  @IsNumber()
-  @IsNotEmpty()
-  userId: number;
-}
