+++ backend/src/chat/dto/socket/socket-join-room.dto.ts
